refactor: slice components
