+++ src/routes/record/UpdatePage.tsx
@@ -8,42 +8,6 @@ import { QUERY_KEYS } from 'utils/queryKeys';
 import { hasReturnUrlField } from './utils/hasReturnUrlField';
 import { Form, FormProps } from './components/Form';
 
-export const CreateRecordPage: FC = () => {
-  const navigate = useNavigate();
-  const { termsRepository } = useServices();
-
-  const {
-    mutateAsync: create,
-    status,
-    error,
-  } = useMutation(termsRepository.create.bind(termsRepository), {
-    onError: e => console.error(e),
-  });
-
-  const navigateBack = () => {
-    navigate('/list');
-  };
-
-  const handleSubmit: FormProps['onSubmit'] = values => {
-    create(values).then(navigateBack).catch(console.error);
-  };
-
-  const handleCancel = () => navigateBack();
-
-  switch (status) {
-    case 'loading':
-      return <SpinnerBox label="Creating ..." />;
-    case 'error':
-      return <p>Error: {error instanceof Error ? error.message : 'Unknown'}</p>;
-    case 'success':
-      return <p>{status}</p>;
-    case 'idle':
-      return <Form onCancel={handleCancel} onSubmit={handleSubmit} />;
-    default:
-      return <p>Unknown status: {status}</p>;
-  }
-};
-
 export const UpdateRecordPage: FC = () => {
   const navigate = useNavigate();
   const location = useLocation();
